Fixed timeLeftInKick bug
timeLeftInKick used to only work with pre-shortened Fifos. It will now
work with all timestamped Fifos. In order to make sure that the Fifo is
always timestamped, canKick was modified to take an extra input,
GameMode, which is used to timestamp the initial Fifo. GameMode was also
added as an input to timeLeftInKick so that it can count only the
remaining moves.