Added function "isBallGoingForGoal"
This new function is used to see if the ball is on course to go into the
opponent's goal. It should be used to get our players out of the way of
the ball so that we can score.
NOTE: This function does not take into account rebounds off of the walls
as it was deemed unnecessary.