Added function "GraphDontBlock"
This new function creates a matrix of the field that should be
multiplied by the matrix used to determine where players should go. This
is only necessary when the ball is heading towards the opponents net so
that we don't accidentally block a goal. This should be used with
"isBallGoingForGoal" function to determine this condition.