Intersection.m modified for expanded use
Intersection now takes in a player's position, a player's type, the
ball's position, and an option offset. The offset is used when the
ball's position is at a point in the future while the player's position
is not. In order to make the inputs more usable, BallPrediction was
modified so that it in turn was also made more usable.