Added new field mapping for moving agents
GraphMoveOut is used to coax players out of our net so that they don't
crowd the goalie and so that they aren't perceived as good targets to
pass to, which was previously leading to inadvertent goals on our own
net.
NB: It may be useful at some point to combine this new matrix with the
inverted field mapping currently used in the code so as to improve
performance. Right now this isn't affecting things too badly.