Static Field Mods. Intelligent interceptions.
Players cannot stand in the opponents' net anymore. Also, a section of
six points in the opponents' net has a value of 0.99999 to help improve
the likelihood of out team's members shooting instead of passing when
the net is wide open.

Players who cannot kick the ball, but are chasing it will go to the
Intersection point of where they can meet up with the ball rather than
just going to the current position of the ball.