New function "PassIntersection.m"
The new function uses a player's fifo and finds
where the ball can intersect with it. If the ball
position passed to the function is a future pos.,
you can pass an offset of how many cycles it has
been shifted. A ball speed is also needed to find
when it will intersect. If none is provided, the
speed will be determined by the current ball
velocity vector.

Other minor changes: Some minor syntax changes and 
bug fixes in PlayerPrediction.m. HLS.m updated to 
reflect syntax changes.